Add files via upload
These are the two files I used to create a project for the Colt Steele Web Developer Boot Camp on Udemy. The project was initially set to utilize Bootstrap 3.3.7, but I went ahead and used the latest Bootstrap 4, along with the newest Fontawesome...it was somewhat of a headache, and I had to utilize Stack Overflow more than a couple of times, but I learned more this way instead of just coding along with Colt.